Evoastra Venture Pvt Ltd/Notebooks/requirements.txt
streamlit
tensorflow
numpy
Pillow
nltk
tqdm